+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.8] - 2019-10-09
+
+### Changed
+- Constant/enum type names are prefixed with parent property if available.
+- Constant/enum item names are prefixed with parent type instead of path. 
+- Schema title is used for struct name instead of path where available.
+
+### Added
+- Optional support for `x-nullable` (Swagger 2.0), `nullable` (OpenAPI 3.0) properties to enable nullability.
+
+### Fixed
+- Type marker stripping from path was affecting regexps too.
+
 ## [0.4.7] - 2019-10-05
 
 ### Added
@@ -44,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Removed unnecessary regexp dependency, #7.
 
+[0.4.8]: https://github.com/swaggest/go-code-builder/compare/v0.4.7...v0.4.8
 [0.4.7]: https://github.com/swaggest/go-code-builder/compare/v0.4.6...v0.4.7
 [0.4.6]: https://github.com/swaggest/go-code-builder/compare/v0.4.5...v0.4.6
 [0.4.5]: https://github.com/swaggest/go-code-builder/compare/v0.4.4...v0.4.5